major naming refactor for move to etddf repository
