+++ .gitignore
@@ -1,2 +1,4 @@
 *.m~
 *.pyc
+venv
+env
