+++ python/ros_wrapper/config/points.yaml
