+++ python/ros_wrapper/docs/images/sys_diagram.png
