+++ python/ros_wrapper/launch/etddf.launch
@@ -5,32 +5,32 @@
 	<node pkg="tf" type="static_transform_publisher" name="world_transform"
         args="0 0 0 0 0 0 world map 100"/>
 	
-	<rosparam command="load" file="$(find offset_etddf)/config/$(arg config_file)" />
+	<rosparam command="load" file="$(find etddf_ros)/config/$(arg config_file)" />
 	
     <!-- simulator node -->
-	<node name="point_sim_controller" pkg="offset_etddf" type="point_sim.py" />
+	<node name="point_sim_controller" pkg="etddf_ros" type="point_sim.py" />
 
     <!-- Launch ET-DDF instances -->
-    <include file="$(find offset_etddf)/launch/agent.launch"> 
+    <include file="$(find etddf_ros)/launch/agent.launch"> 
         <arg name="id" value="0" />
         <arg name="planner" value="true" />
         <arg name="sensors" value="true" />
         <!-- <arg name="log_level" value="DEBUG" /> -->
     </include>
 
-    <include file="$(find offset_etddf)/launch/agent.launch"> 
+    <include file="$(find etddf_ros)/launch/agent.launch"> 
         <arg name="id" value="1" />
         <arg name="planner" value="true" />
         <arg name="sensors" value="true" />
     </include>
 
-    <include file="$(find offset_etddf)/launch/agent.launch"> 
+    <include file="$(find etddf_ros)/launch/agent.launch"> 
         <arg name="id" value="2" />
         <arg name="planner" value="true" />
         <arg name="sensors" value="true" />
     </include>
 
-    <include file="$(find offset_etddf)/launch/agent.launch"> 
+    <include file="$(find etddf_ros)/launch/agent.launch"> 
         <arg name="id" value="3" />
         <arg name="planner" value="true" />
         <arg name="sensors" value="true" />
