+++ python/ros_wrapper/msg/gpsMeasurement.msg
