+++ python/setup.py
@@ -4,11 +4,11 @@ from distutils.core import setup
 # from catkin_pkg.python_setup import generate_distutils_setup
 
 setup(
-    name='offset',
-    description='offset project event-triggered distributed data fusion',
+    name='etddf',
+    description='event-triggered decentralized data fusion package',
     author='Ian Loefgren',
     author_email='ian.loefgren@colorado.edu',
-    packages=['offset','offset.filters']
+    packages=['etddf','etddf.filters']
 )
 
 # setup_args = generate_distutils_setup(
